+++ client/src/app/apiService.service.ts
@@ -2,14 +2,14 @@ import { Http } from '@angular/http';
 import { Injectable } from '@angular/core';
 
 @Injectable()
-export class HomeService {
+export class ApiService {
+    messages = [];
 
     constructor(private http: Http) {}
 
     sendLocationData(locationData) {
-        this.http.post('http://localhost:8100/', locationData).subscribe(res => {
-            console.log(locationData);
-            console.log(res);
+        this.http.get('http://localhost:3000/posts', locationData).subscribe(res => {
+            console.log(res.json());
         });
     }
-}
+}
