+++ client/src/app/app.module.ts
@@ -1,3 +1,4 @@
+import { ApiService } from './apiService.service';
 import { BrowserModule } from '@angular/platform-browser';
 import { ErrorHandler, NgModule } from '@angular/core';
 import { IonicApp, IonicErrorHandler, IonicModule } from 'ionic-angular';
@@ -5,14 +6,12 @@ import { AgmCoreModule } from '@agm/core';
 import { HttpModule } from '@angular/http';
 
 import { MyApp } from './app.component';
-import { HomePage } from '../pages/home/home.component';
+import { HomePage } from '../pages/home/home';
 import { ListPage } from '../pages/list/list';
 
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
 import { Geolocation } from '@ionic-native/geolocation';
-
-import { HomeService } from '../pages/home/home.service';
 
 @NgModule({
   declarations: [
@@ -37,7 +36,7 @@ import { HomeService } from '../pages/home/home.service';
   providers: [
     StatusBar,
     SplashScreen,
-    HomeService,
+    ApiService,
     Geolocation,
     {provide: ErrorHandler, useClass: IonicErrorHandler}
   ]
